+++ lib/src/enums/player_state.dart
@@ -1,4 +1,4 @@
-// Copyright 2019 Sarbagya Dhaubanjar. All rights reserved.
+// Copyright 2020 Sarbagya Dhaubanjar. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -12,9 +12,7 @@ enum PlayerState {
   /// Denotes state when player loads first video.
   unStarted,
 
-  /// Denotes state when player ends playing a video.
-  ///
-  /// Fires only once after video ends and switches to [PlayerState.stopped].
+  /// Denotes state when player has ended playing a video.
   ended,
 
   /// Denotes state when player is playing video.
@@ -28,7 +26,4 @@ enum PlayerState {
 
   /// Denotes state when player loads video and is ready to be played.
   cued,
-
-  /// Denotes state when player has stopped playing videos.
-  stopped,
 }