+++ lib/src/widgets/duration_widgets.dart
@@ -1,4 +1,4 @@
-// Copyright 2019 Sarbagya Dhaubanjar. All rights reserved.
+// Copyright 2020 Sarbagya Dhaubanjar. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -107,7 +107,7 @@ class _RemainingDurationState extends State<RemainingDuration> {
   Widget build(BuildContext context) {
     return Text(
       "- ${durationFormatter(
-        (_controller.value.duration?.inMilliseconds ?? 0) -
+        (_controller.metadata.duration?.inMilliseconds ?? 0) -
             (_controller.value.position?.inMilliseconds ?? 0),
       )}",
       style: TextStyle(