+++ lib/src/widgets/progress_bar.dart
@@ -1,4 +1,4 @@
-// Copyright 2019 Sarbagya Dhaubanjar. All rights reserved.
+// Copyright 2020 Sarbagya Dhaubanjar. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -89,7 +89,7 @@ class _ProgressBarState extends State<ProgressBar> {
   }
 
   void positionListener() {
-    var _totalDuration = _controller.value.duration?.inMilliseconds;
+    var _totalDuration = _controller.metadata.duration?.inMilliseconds;
     if (mounted && !_totalDuration.isNaN && _totalDuration != 0) {
       setState(() {
         _playedValue =
@@ -117,13 +117,13 @@ class _ProgressBarState extends State<ProgressBar> {
     _touchPoint = box.globalToLocal(globalPosition);
     _checkTouchPoint();
     final relative = _touchPoint.dx / box.size.width;
-    _position = _controller.value.duration * relative;
+    _position = _controller.metadata.duration * relative;
     _controller.seekTo(_position, allowSeekAhead: false);
   }
 
   void _dragEndActions() {
     _controller.updateValue(
-      _controller.value.copyWith(showControls: false, isDragging: false),
+      _controller.value.copyWith(isControlsVisible: false, isDragging: false),
     );
     _controller.seekTo(_position, allowSeekAhead: true);
     setState(() {
@@ -136,7 +136,7 @@ class _ProgressBarState extends State<ProgressBar> {
     return GestureDetector(
       onHorizontalDragDown: (details) {
         _controller.updateValue(
-          _controller.value.copyWith(showControls: true, isDragging: true),
+          _controller.value.copyWith(isControlsVisible: true, isDragging: true),
         );
         _seekToRelativePosition(details.globalPosition);
         setState(() {
@@ -237,7 +237,7 @@ class _ProgressBarPainter extends CustomPainter {
 
     if (touchDown) {
       handlePaint.color = _handleColor.withOpacity(0.4);
-      canvas.drawCircle(progressPoint, handleRadius * 2, handlePaint);
+      canvas.drawCircle(progressPoint, handleRadius * 3, handlePaint);
     }
 
     handlePaint.color = _handleColor;